+++ src/hl/plist/file_access.rs
@@ -1026,6 +1026,11 @@ impl FileAccessBuilder {
         Ok(builder)
     }
 
+    /// Sets the file close degree
+    ///
+    /// If called with `FileCloseDegree::Strong`, the programmer is responsible
+    /// for closing all items before closing the file. Failure to do so might
+    /// invalidate newly created objects.
     pub fn fclose_degree(&mut self, fc_degree: FileCloseDegree) -> &mut Self {
         self.fclose_degree = Some(fc_degree);
         self
@@ -1374,6 +1379,8 @@ impl FileAccessBuilder {
         if let Some(v) = self.chunk_cache {
             h5try!(H5Pset_cache(id, 0, v.nslots as _, v.nbytes as _, v.w0 as _));
         }
+        // The default is to use CLOSE_SEMI or CLOSE_WEAK, depending on VFL driver.
+        // Both of these are unproblematic for our ownership
         if let Some(v) = self.fclose_degree {
             h5try!(H5Pset_fclose_degree(id, v.into()));
         }
@@ -1535,6 +1542,9 @@ impl FileAccess {
                 files.push(MultiFile::new(&string_from_cstr(name), addr as _));
             }
             *layout.get_mut(i - 1) = 0xff - mapping[j];
+        }
+        for &memb_name in &memb_name {
+            crate::util::h5_free_memory(memb_name as *mut _);
         }
         let relax = relax > 0;
         let drv = MultiDriver { files, layout, relax };