+++ CHANGELOG.md
@@ -51,7 +51,9 @@
 - Added support for creating external links on a `Group` with `link_external`.
 - Added `Location` methods: `get_info`, `get_info_by_name`, `loc_type`, and `open_by_token`.
 - Added `Group` methods: `iter_visit`, `iter_visit_default`, `get_all_of_type`, `datasets`, `groups`, and `named_datatypes`.
-  
+- Added `Debug` for `Handle`.
+- Added method `try_borrow` on `Handle` for not taking ownership of an HDF5 object.
+
  ### Changed
 
 - Required Rust compiler version is now `1.51`.
@@ -78,6 +80,12 @@
 - `silence_errors` now work globally and will not be reset on dropping `SilenceErrors`.
 - Errors are not expanded when encountered, but only when being used for printing or by
   the library user.
+- Handles to `hdf5` identifiers are no longer tracked via a registry and is instead handled by stricter semantics of ownership.
+- The handle to a `File` will not close all objects in a `File` when dropped, but instead uses a weak file close degree. For the old behaviour see `FileCloseDegree::Strong`.
+
+### Fixed
+
+- A memory leak of handles has been identified and fixed.
 
 ## 0.7.1